Update block_bots.conf
Removed "RequireAll" block from here. It was conflicting with other RequireAny or RequireAll blocks on the VirtualHost that includes it. The problem was that there it was written: "Require all granted", so, if you added afterwards a block restrinting ips, it won't be excecuted because the access was already granted. For a solution see the comments on this file.

I also removed the comments for Apache 2.2. I don't know how to port it.